Did four things:
a. Added the line "invert net" at the end of the scripts
b. ... and hence also changed names from latin2syllabics to syllabics2latin and vice versa.
c. moved the decapitalisation rule from file to file so that it stays with the latin2syllabics file
d. commented out a zero to hyphen rule, it did not work after the inversion. Consider reintroducing it in the opposite file.

With this setup (and Antti's rules), I managed to make a roundtrip conversion syll > lat > syll with a small number of errors (still enough that the issue is not settled yet). Please have a look, and consider whether this is an improvement or not.



git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/chr@97408 c7155fb1-f0a7-4240-a2fc-2600b6f42f90